fix versions for spark jar path
In the bootstrap.sh script, the following JAR file had changes
in the versions that are retrieved from the web.

mongo-java-driver (3.4.2 -> 3.6.1)
elasticsearch-spark (20_2.10-5.2.1 -> 20_2.11-6.1.2)

However, the echo command that writes JAR file names to
the spark configuration file still has the old versions.
This commit updates those to prevent exceptions being
thrown when pyspark is being launched.

Might be worth setting some variables in the script that
track what versions are being used so that having to
edit the script multiple places in the future is avoided.