add disk resize parameter to box configuration, add port 9200
This commit introduces a potentially controversial change.
In testing with the latest ubuntu/bionic64 Vagrant box

    (virtualbox, 20181126.0.0)

with the latest updates to bootstrap.sh, I found that the
default disk size, 10 GB is no longer sufficient. The
OS will run out of disk space before the provisioning
process completes.  Vagrant does not have any built-in
way to specify a disk resize, but there is a plug-in.

https://github.com/sprotheroe/vagrant-disksize

An example of its usage is provided here:

https://nmrony.info/change-disk-size-of-a-vagrant-box/

Setting config.disksize.size to '20GB' suffices.  You might
be able to get away with '15GB', but I am not far enough
down the road in the book to know if this is safe.

I also noticed that the test query to elasticsearch in
boostrap.sh fails with a connection refused error.  I wasn't
sure if this was being run on the host or the guest VM, so
I tried adding 9200 to the list of port mappings from
host to guest, but that did not fix the issue.  Went ahead
and left the setting in place because it allows you to
issue the curl commands in the book covered in Chapter 2
without having to vagrant ssh in.